♻️ refactor(auth): 统一替换用户类型魔法数字为常量引用
- 在 Auth Handler 中使用常量替代用户类型魔法数字（0=会员，1=管理员）
- 在 Member Auth Handler 中统一用户类型常量引用
- 优化 AuthService：使用常量提升代码可读性和维护性
- 完善 MemberAuthService：规范用户类型处理
- 改进 MemberUserService：统一常量约定
- 提升代码质量和可维护性

+++ internal/service/member/user.go
@@ -119,7 +119,7 @@ func (s *MemberUserService) UpdateUser(ctx context.Context, id int64, req *req.A
 // UpdateUserMobile 修改用户手机
 func (s *MemberUserService) UpdateUserMobile(ctx context.Context, id int64, req *req.AppMemberUserUpdateMobileReq) error {
 	// 1. 校验验证码
-	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, req.Scene, req.Code); err != nil {
+	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, int32(req.Scene), req.Code); err != nil {
 		return err
 	}
 
@@ -141,7 +141,7 @@ func (s *MemberUserService) UpdateUserMobile(ctx context.Context, id int64, req 
 func (s *MemberUserService) ResetUserPassword(ctx context.Context, req *req.AppMemberUserResetPasswordReq) error {
 	// 1. 校验验证码 (场景: 重置密码)
 	// TODO: Replace magic number with Enum. MEMBER_RESET_PASSWORD = 4
-	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, 4, req.Code); err != nil {
+	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, int32(4), req.Code); err != nil {
 		return err
 	}
 
@@ -166,7 +166,7 @@ func (s *MemberUserService) ResetUserPassword(ctx context.Context, req *req.AppM
 // UpdateUserPassword 修改用户密码
 func (s *MemberUserService) UpdateUserPassword(ctx context.Context, id int64, req *req.AppMemberUserUpdatePasswordReq) error {
 	// 1. 校验验证码
-	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, req.Scene, req.Code); err != nil {
+	if err := s.smsCodeSvc.ValidateSmsCode(ctx, req.Mobile, int32(req.Scene), req.Code); err != nil {
 		return err
 	}
